change code structure
