+++ include/micros_rtt/common.h
@@ -16,6 +16,7 @@ namespace micros_rtt
 {
 
 enum FlowStatus {NoData = 0, OldData = 1, NewData = 2};
+
  
 }
